Moved exception class files to a subdirectory.
