--- lib/ElementNotSelectableException.php
@@ -1,7 +1,0 @@
-<?php
-/*
- * Error code 14 as defined in the json wired protocol
- */
-
-class ElementNotSelectableException extends WebDriverException {}
-
