+++ lib/__init__.php
@@ -116,37 +116,37 @@ require_once('interactions/touch/WebDriverTapAction.php');
 require_once('interactions/touch/WebDriverUpAction.php');
 
 // exceptions
-require_once('IndexOutOfBoundsException.php');
-require_once('NoCollectionException.php');
-require_once('NoStringException.php');
-require_once('NoStringLengthException.php');
-require_once('NoStringWrapperException.php');
-require_once('NoSuchDriverException.php');
-require_once('NoSuchElementException.php');
-require_once('NoSuchFrameException.php');
-require_once('UnknownCommandException.php');
-require_once('StaleElementReferenceException.php');
-require_once('ElementNotVisibleException.php');
-require_once('InvalidElementStateException.php');
-require_once('UnknownServerException.php');
-require_once('ExpectedException.php');
-require_once('ElementNotSelectableException.php');
-require_once('NoSuchDocumentException.php');
-require_once('UnexpectedJavascriptException.php');
-require_once('NoScriptResultException.php');
-require_once('XPathLookupException.php');
-require_once('NoSuchCollectionException.php');
-require_once('TimeOutException.php');
-require_once('NullPointerException.php');
-require_once('NoSuchWindowException.php');
-require_once('InvalidCookieDomainException.php');
-require_once('UnableToSetCookieException.php');
-require_once('UnexpectedAlertOpenException.php');
-require_once('NoAlertOpenException.php');
-require_once('ScriptTimeoutException.php');
-require_once('InvalidCoordinatesException.php');
-require_once('IMENotAvailableException.php');
-require_once('IMEEngineActivationFailedException.php');
-require_once('InvalidSelectorException.php');
-require_once('SessionNotCreatedException.php');
-require_once('MoveTargetOutOfBoundsException.php');
+require_once('exceptions/IndexOutOfBoundsException.php');
+require_once('exceptions/NoCollectionException.php');
+require_once('exceptions/NoStringException.php');
+require_once('exceptions/NoStringLengthException.php');
+require_once('exceptions/NoStringWrapperException.php');
+require_once('exceptions/NoSuchDriverException.php');
+require_once('exceptions/NoSuchElementException.php');
+require_once('exceptions/NoSuchFrameException.php');
+require_once('exceptions/UnknownCommandException.php');
+require_once('exceptions/StaleElementReferenceException.php');
+require_once('exceptions/ElementNotVisibleException.php');
+require_once('exceptions/InvalidElementStateException.php');
+require_once('exceptions/UnknownServerException.php');
+require_once('exceptions/ExpectedException.php');
+require_once('exceptions/ElementNotSelectableException.php');
+require_once('exceptions/NoSuchDocumentException.php');
+require_once('exceptions/UnexpectedJavascriptException.php');
+require_once('exceptions/NoScriptResultException.php');
+require_once('exceptions/XPathLookupException.php');
+require_once('exceptions/NoSuchCollectionException.php');
+require_once('exceptions/TimeOutException.php');
+require_once('exceptions/NullPointerException.php');
+require_once('exceptions/NoSuchWindowException.php');
+require_once('exceptions/InvalidCookieDomainException.php');
+require_once('exceptions/UnableToSetCookieException.php');
+require_once('exceptions/UnexpectedAlertOpenException.php');
+require_once('exceptions/NoAlertOpenException.php');
+require_once('exceptions/ScriptTimeoutException.php');
+require_once('exceptions/InvalidCoordinatesException.php');
+require_once('exceptions/IMENotAvailableException.php');
+require_once('exceptions/IMEEngineActivationFailedException.php');
+require_once('exceptions/InvalidSelectorException.php');
+require_once('exceptions/SessionNotCreatedException.php');
+require_once('exceptions/MoveTargetOutOfBoundsException.php');
