+++ lib/exceptions/ElementNotSelectableException.php
@@ -1,0 +1,7 @@
+<?php
+/*
+ * Error code 14 as defined in the json wired protocol
+ */
+
+class ElementNotSelectableWebDriverException extends WebDriverException {}
+
