+++ lib/exceptions/ExpectedException.php
