+++ lib/exceptions/IMENotAvailableException.php
