+++ lib/exceptions/InvalidCoordinatesException.php
