+++ lib/exceptions/InvalidSelectorException.php
