+++ lib/exceptions/NoAlertOpenException.php
