+++ lib/exceptions/NoScriptResultException.php
