+++ lib/exceptions/NoStringException.php
