+++ lib/exceptions/NoStringLengthException.php
