+++ lib/exceptions/NoStringWrapperException.php
