+++ lib/exceptions/NoSuchFrameException.php
