+++ lib/exceptions/SessionNotCreatedException.php
