+++ lib/exceptions/UnableToSetCookieException.php
