+++ lib/exceptions/UnexpectedJavascriptException.php
